move shared resource to different file
